Significant update
Removed "band aid" code, fixed many bugs, updated user interaction model, performance improvements for tstats and web interaction

+++ tquote.py
@@ -1,16 +1,22 @@
 #!/usr/bin/python3
 
 import random
+from os import listdir, mkdir
+from os.path import isdir
+import asyncio
+import aiohttp
+import aiofiles
+from pathlib import Path
 
 import pendulum
-from sqlalchemy import create_engine, and_, func, update, select, MetaData, Table, Column, Integer, String 
+from sqlalchemy import and_, func, update, select, MetaData, Table, Column, Integer, String 
 
-from tutil import debug, fetchFile, is_admin
+from tutil import debug, fetchFile, is_admin, incrementUsage
+from constants import DEFAULT_DIR, ENGINE
 
 			
 def setup():
-	global engine, meta, Quotes, Lore, Config
-	engine = create_engine('sqlite:///./log/quotes.db', echo = False)
+	global meta, Quotes, Lore, Config
 	meta = MetaData()
 	Config = Table(
 		'config', meta,
@@ -31,6 +37,7 @@ def setup():
 		Column('date', String),
 		Column('guild', String),
 		Column('guild_name', String),
+		Column('embed', String),
 		)	
 	Lore = Table(
 		'famLore', meta,
@@ -40,37 +47,48 @@ def setup():
 		Column('date', String),
 		Column('guild', String),
 		)
-	meta.create_all(engine)
+	meta.create_all(ENGINE)
 	print('[+] End Quotes Setup')
 
 
 def helper(message):
-	args = message.content.split()
-		
+	text = message.content
+	for each in message.mentions:
+		text = text.replace('<@!{}>'.format(each.id), each.name)
+	args = text.split()
+	guild = message.guild.id
+
+	#Lore	
 	if args[0] == '!lore':
+		incrementUsage(message.guild, 'lore')
 		if len(args) > 1:
 			if args[1] == 'add' and is_admin(message.author):
-				return insertQuote(Lore, message)
+				return insertQuote(message, Lore)
 			elif args[1] == 'help' and is_admin(message.author):
 				return getHelp(message.author).split('@LORE')[1]
 			else: 
-				return getQuote(Lore, message.guild.id, ' '.join(args[1:]))
+				return getQuote(guild, Lore, ' '.join(args[1:]))
 		else:
-			return getQuote(Lore, message.guild.id)
-			
+			return getQuote(guild, Lore)
+
+	#Quote with options		
 	elif len(args) > 1:
+		incrementUsage(message.guild, 'quote')
 		if args[1] == 'help':
 			return getHelp(message.author).split('@LORE')[0]
 		else:
-			return getQuote(Quotes, message.guild.id, ' '.join(args[1:]))
+			return getQuote(guild, Quotes, ' '.join(args[1:]))
+	
+	#Any random quote
 	else:
-		return getQuote(Quotes, message.guild.id)
+		incrementUsage(message.guild, 'quote')
+		return getQuote(guild, Quotes)
 	
 
-def insertQuote(Table, message):
-	if Table == None:
+def insertQuote(message, Table):
+	"""Insert a quote to the database"""
+	if Table is None:
 		Table = Quotes
-	
 	config = load_config(message.guild.id)
 	if config:
 		server_locale = config[2]
@@ -78,16 +96,19 @@ def insertQuote(Table, message):
 	else:
 		server_locale = 'Asia/Tokyo'
 		stm = 'Added: "{}" by {} on {}'
-	date = pendulum.now(tz=server_locale).to_day_datetime_string()
 	
+	#Supress any user or role mentions
 	text = message.content
 	for each in message.mentions:
 		text = text.replace('<@!{}>'.format(each.id), each.name)
 	for each in message.role_mentions:
 		text = text.replace('<@&{}>'.format(each.id), each.name)
-	args = text.split()	
 	
-	with engine.connect() as conn:
+	args = text.split()	
+	embed = str(message.attachments[0].url) if message.attachments else None
+	date = pendulum.now(tz=server_locale).to_day_datetime_string()
+	
+	with ENGINE.connect() as conn:
 		if Table.name == 'famQuotes':
 			ins = Table.insert().values(
 				id = message.id,
@@ -96,6 +117,7 @@ def insertQuote(Table, message):
 				date = date,
 				guild = str(message.guild.id),
 				guild_name = message.guild.name,
+				embed = embed,
 			)
 			conn.execute(ins)
 			return stm.format(text, message.author.name, date)
@@ -112,21 +134,20 @@ def insertQuote(Table, message):
 			return stm.format(' '.join(args[3:]), args[2], date)
 
 
-def getQuote(Table, guild_, id_=None):
-	if Table == None:
-		Table = Quotes
-	if id_:
+def getQuote(guild, Table, username=None):
+	"""Retrive a quote from the database"""
+	if username:
 		select_st = select([Table]).where(and_(
-			Table.c.name == id_,
-			Table.c.guild == guild_)).order_by(func.random())
+			Table.c.name == username,
+			Table.c.guild == guild)).order_by(func.random())
 	else:
 		select_st = select([Table]).where(
-			Table.c.guild == guild_).order_by(func.random())
+			Table.c.guild == guild).order_by(func.random())
 			
-	with engine.connect() as conn:
+	with ENGINE.connect() as conn:
 		result = conn.execute(select_st).fetchone()
 		if result:
-			config = load_config(guild_)
+			config = load_config(guild)
 			if config:			
 				if(Table.name == 'famQuotes'):
 					stm = config[4].replace('\\n', '\n')
@@ -137,33 +158,93 @@ def getQuote(Table, guild_, id_=None):
 					stm = '{}\n ---{} on {}'
 				elif(Table.name == 'famLore'):
 					stm = '{}\n ---Scribed by the Lore Master {}, on the blessed day of {}'
-			#result[1]: author, [2]: quote, [3]: date
+			#Check if there is an attached img or file to send as well
+			if len(result) > 6 and result[6]:
+				stm = stm + '\n' + result[6]
+			#Result fields translate as
+			#[1]: author, [2]: quote, [3]: date, [6]: embed url
 			return stm.format(result[2], result[1], result[3])
 
 
-def deleteQuote(guild_, id_):
-	with engine.connect() as conn:
+def deleteQuote(guild, msg_id):
+	"""Remove a quote from the database"""
+	with ENGINE.connect() as conn:
 		for Table in {Quotes, Lore}:
-			ins = Table.delete().where(and_(
-				Table.c.id == id_,
-				Table.c.guild == guild_))
-			conn.execute(ins)
-	if not checkExists(guild_, id_):
-		return 'Deleted quote {}'.format(id_)
-	return 'Error'
+			select_st = select([Table]).where(and_(
+				Table.c.id == msg_id,
+				Table.c.guild == guild
+			))
+			result = conn.execute(select_st).fetchone()
+			if result:
+				quote = '{}\n ---{} on {}'.format(result[2], result[1], result[3])
+
+				ins = Table.delete().where(and_(
+					Table.c.id == msg_id,
+					Table.c.guild == guild
+					))
+				conn.execute(ins)
+
+				#Sanity check
+				if not checkExists(guild, msg_id):
+					return 'Deleted quote:\n`{}`'.format(quote)
 
 
-def checkExists(guild_, id_):
-	with engine.connect() as conn:
+def checkExists(guild, msg_id):
+	"""Ensure that we do not add the same message to the database multiple times"""
+	with ENGINE.connect() as conn:
 		select_st = select([Quotes]).where(and_(
-			Quotes.c.id == id_,
-			Quotes.c.guild == guild_))
+			Quotes.c.id == msg_id,
+			Quotes.c.guild == guild))
 		if conn.execute(select_st).fetchall(): return True
 	return False	
+
+
+def getReact(message, nsfw):
+	"""Get a random gif file from ./emotes or the servers folder"""
+	incrementUsage(message.guild, 'gif')
+
+	guild = message.guild.id
+	reacts = ['{}/emotes/{}'.format(DEFAULT_DIR, each) for each in listdir('./emotes')]
+	
+	if isdir('{}/emotes/{}'.format(DEFAULT_DIR, guild)):
+		reacts.extend(['{}/emotes/{}/{}'.format(DEFAULT_DIR, guild, each) for each in listdir('./emotes/{}'.format(guild))])
+	
+	if nsfw and isdir('{}/emotes/{}/nsfw'.format(DEFAULT_DIR, guild)):
+		reacts.extend(['{}/emotes/{}/nsfw/{}'.format(DEFAULT_DIR, guild, each) for each in listdir('./emotes/{}/nsfw'.format(guild))])
+
+	return random.choice(reacts)
+
+
+async def fetchReact(message):
+	"""Save a gif a user added with !gif add"""
+	incrementUsage(message.guild, 'gif')
+	url = str(message.attachments[0].url)
+	ext = str(url.split('.')[-1].lower())
+    
+	if ext != 'gif':
+		return 'File must be a gif'
+        
+	fileName = str(url.split('/')[-1])    
+	nsfw = True if 'nsfw' in message.content.lower() else False
+    
+	if nsfw:
+		Path('{}/emotes/{}/nsfw'.format(DEFAULT_DIR, message.guild.id)).mkdir(parents=True, exist_ok=True)    
+		filePath = '{}/emotes/{}/nsfw/{}'.format(DEFAULT_DIR, message.guild.id, fileName)
+	else:
+		Path('{}/emotes/{}'.format(DEFAULT_DIR, message.guild.id)).mkdir(parents=True, exist_ok=True)    
+		filePath = '{}/emotes/{}/{}'.format(DEFAULT_DIR, message.guild.id, fileName)
+    
+	async with aiohttp.ClientSession() as session:
+		async with session.get(url) as resp:
+			if resp.status == 200:
+				f = await aiofiles.open(filePath, mode='wb')
+				await f.write(await resp.read())
+				await f.close()
+				print("[+] Saved: {}".format(filePath))  
 	
 	
 def load_config(guild):
-	with engine.connect() as conn:
+	with ENGINE.connect() as conn:
 		select_st = select([Config]).where(Config.c.id == guild)
 		result = conn.execute(select_st).fetchone()
 	if result:
